requirements.txt
pandas
streamlit
gitpython
nltk
allennlp==1.0.0    
allennlp-models==1.0.0
transformers==3.0.0